Relocated generic project link route
